Move Schemas to base module
