+++ tests/test_schema.py
@@ -8,101 +8,7 @@ from marshmallow import validate, ValidationError
 from marshmallow_jsonapi import Schema, fields
 from marshmallow_jsonapi.exceptions import IncorrectTypeError
 from marshmallow_jsonapi.utils import _MARSHMALLOW_VERSION_INFO
-from tests.base import unpack
-
-
-class AuthorSchema(Schema):
-    id = fields.Str()
-    first_name = fields.Str(required=True)
-    last_name = fields.Str(required=True)
-    password = fields.Str(load_only=True, validate=validate.Length(6))
-    twitter = fields.Str()
-
-    def get_top_level_links(self, data, many):
-        if many:
-            self_link = '/authors/'
-        else:
-            self_link = '/authors/{}'.format(data['id'])
-        return {'self': self_link}
-
-    class Meta:
-        type_ = 'people'
-        strict = True  # for marshmallow 2
-
-
-class PostSchema(Schema):
-    id = fields.Str()
-    post_title = fields.Str(attribute='title', dump_to='title', data_key='title')
-
-    author = fields.Relationship(
-        'http://test.test/posts/{id}/author/',
-        related_url_kwargs={'id': '<id>'},
-        schema=AuthorSchema, many=False,
-        type_='people'
-    )
-
-    post_comments = fields.Relationship(
-        'http://test.test/posts/{id}/comments/',
-        related_url_kwargs={'id': '<id>'},
-        attribute='comments',
-        load_from='post-comments', dump_to='post-comments',
-        data_key='post-comments',
-        schema='CommentSchema', many=True,
-        type_='comments'
-    )
-
-    post_keywords = fields.Relationship(
-        'http://test.test/posts/{id}/keywords/',
-        related_url_kwargs={'id': '<id>'},
-        attribute='keywords', dump_to='post-keywords',
-        data_key='post-keywords',
-        schema='KeywordSchema', many=True,
-        type_='keywords'
-    )
-
-    class Meta:
-        type_ = 'posts'
-        strict = True
-
-
-class CommentSchema(Schema):
-    id = fields.Str()
-    body = fields.Str(required=True)
-
-    author = fields.Relationship(
-        'http://test.test/comments/{id}/author/',
-        related_url_kwargs={'id': '<id>'},
-        schema=AuthorSchema, many=False
-    )
-
-    class Meta:
-        type_ = 'comments'
-        strict = True
-
-
-class KeywordSchema(Schema):
-    id = fields.Str()
-    keyword = fields.Str(required=True)
-
-    def get_attribute(self, attr, obj, default):
-        if _MARSHMALLOW_VERSION_INFO[0] >= 3:
-            if obj == 'id':
-                return md5(super(Schema, self)
-                           .get_attribute(attr, 'keyword', default)
-                           .encode('utf-8')).hexdigest()
-            else:
-                return super(Schema, self).get_attribute(attr, obj, default)
-        else:
-            if attr == 'id':
-                return md5(super(Schema, self)
-                           .get_attribute('keyword', obj, default)
-                           .encode('utf-8')).hexdigest()
-            else:
-                return super(Schema, self).get_attribute(attr, obj, default)
-
-    class Meta:
-        type_ = 'keywords'
-        strict = True
+from tests.base import unpack, AuthorSchema, CommentSchema, PostSchema, PolygonSchema, ArticleSchema
 
 
 def test_type_is_required():
@@ -818,21 +724,6 @@ class TestAutoSelfUrls:
         assert 'links' not in data
 
 
-class PolygonSchema(Schema):
-    id = fields.Integer(as_string=True)
-    sides = fields.Integer()
-    # This is an attribute that uses the 'meta' key: /data/attributes/meta
-    meta = fields.String()
-    # This is the document's top level meta object: /meta
-    document_meta = fields.DocumentMeta()
-    # This is the resource object's meta object: /data/meta
-    resource_meta = fields.ResourceMeta()
-
-    class Meta:
-        type_ = 'shapes'
-        strict = True
-
-
 class TestMeta(object):
     shape = {
         'id': 1,
@@ -902,19 +793,6 @@ class TestMeta(object):
         assert second['document_meta'] == self.shapes[1]['document_meta']
 
 
-class ArticleSchema(Schema):
-    id = fields.Integer()
-    body = fields.String()
-    author = fields.Relationship(
-        dump_only=False, include_resource_linkage=True, many=False, type_='people')
-    comments = fields.Relationship(
-        dump_only=False, include_resource_linkage=True, many=True, type_='comments')
-
-    class Meta:
-        type_ = 'articles'
-        strict = True
-
-
 def assert_relationship_error(pointer, errors):
     """Walk through the dictionary and determine if a specific
     relationship pointer exists
@@ -927,7 +805,6 @@ def assert_relationship_error(pointer, errors):
 
 
 class TestRelationshipLoading(object):
-
     article = {
         'data': {
             "id": "1",
